// lib/api/schemas/update-user-by-admin-dto.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export const updateUserByAdminDtoRoleEnum = {
	ADMIN: "ADMIN",
	USER: "USER",
} as const;

export type UpdateUserByAdminDtoRoleEnumKey =
	(typeof updateUserByAdminDtoRoleEnum)[keyof typeof updateUserByAdminDtoRoleEnum];

export type UpdateUserByAdminDto = {
	/**
	 * @type string | undefined
	 */
	name?: string;
	/**
	 * @type string | undefined
	 */
	email?: string;
	/**
	 * @type string | undefined
	 */
	role?: UpdateUserByAdminDtoRoleEnumKey;
};